+++ Cargo.toml
@@ -4,5 +4,26 @@ version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 
+[lib]
+crate-type = ["lib", "cdylib", "staticlib"]
+
 [dependencies]
 bevy = "0.6"
+
+# Android configuration
+
+[package.metadata.android]
+build_targets = ["aarch64-linux-android", "armv7-linux-androideabi"]
+assets = "assets"
+resources = "assets/res"
+
+[package.metadata.android.sdk]
+target_sdk_version = 29
+min_sdk_version = 16
+
+[package.metadata.android.application]
+debuggable = true
+icon = "@mipmap/icon"
+
+[package.metadata.android.application.activity]
+orientation = "portrait"
