+++ apple/bevy_github_ci_template.xcodeproj/.gitignore
@@ -1,0 +1,2 @@
+xcuserdata
+project.xcworkspace
