+++ apple/src/bindings.h
@@ -1,0 +1,1 @@
+void main_rs(void);
