+++ apple/src/main.m
@@ -1,0 +1,6 @@
+#import "bindings.h"
+
+int main() {
+    main_rs();
+    return 0;
+}
