+++ src/lib.rs
@@ -1,0 +1,17 @@
+use bevy::prelude::*;
+
+#[bevy_main]
+pub fn main() {
+    App::new()
+        .add_plugins(DefaultPlugins)
+        .add_startup_system(setup)
+        .run();
+}
+
+fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
+    commands.spawn_bundle(OrthographicCameraBundle::new_2d());
+    commands.spawn_bundle(SpriteBundle {
+        texture: asset_server.load("icon.png"),
+        ..Default::default()
+    });
+}
